Revert "camera: wip"
This reverts commit 7bc2114852cccc5e4a8db0c3c4cb02083634460d.

Revert "bacon: camera: Updates from CAF"

This reverts commit 1f55593626e3c671b0986fa69478c62941da438c.

+++ camera/QCamera2/HAL/QCameraAllocator.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2012-2014, The Linux Foundataion. All rights reserved.
+/* Copyright (c) 2012, The Linux Foundataion. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -42,9 +42,13 @@ class QCameraHeapMemory;
 class QCameraAllocator {
 public:
     virtual QCameraMemory *allocateStreamBuf(cam_stream_type_t stream_type,
-            size_t size, int stride, int scanline, uint8_t &bufferCnt) = 0;
+                                             int size,
+                                             int stride,
+                                             int scanline,
+                                             uint8_t &bufferCnt) = 0;
     virtual int32_t allocateMoreStreamBuf(QCameraMemory *mem_obj,
-            size_t size, uint8_t &bufferCnt) = 0;
+                                          int size,
+                                          uint8_t &bufferCnt) = 0;
     virtual QCameraHeapMemory *allocateStreamInfoBuf(cam_stream_type_t stream_type) = 0;
     virtual ~QCameraAllocator() {}
 };
